Fixed some small issued. Moved some of the parameters to the translation model. Propagated model parameters up to the level of the model builders. Now it is time to take the model parameters into account. Also, the TM and RM probabilities are not in the log format, so we need to convert (?!)

+++ inc/server/lm/lm_consts.hpp
@@ -44,7 +44,7 @@ namespace uva {
                 namespace lm {
 
                     //This typedef if used in the tries in order to specify the type of the N-gram level NcontextHash
-                    typedef uint8_t TModelLevel;
+                    typedef uint16_t TModelLevel;
 
                     //The type used for storing log probabilities and back-off values
                     typedef float TLogProbBackOff;
